+++ core/src/test/java/org/elasticsearch/search/sort/SortBuilderTests.java
@@ -57,6 +57,7 @@ public class SortBuilderTests extends ESTestCase {
     @AfterClass
     public static void afterClass() throws Exception {
         namedWriteableRegistry = null;
+        indicesQueriesRegistry = null;
     }
 
     /**
